backend/lambda_trades/requirements.txt
sqlalchemy
pymysql  # For MySQL (RDS)
bcrypt
pyjwt
boto3